python-rag-groq-backup/requirements.txt
# Core dependencies for Groq RAG Service
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Groq AI API
groq==0.4.1

# Vector database and embeddings
chromadb==0.4.18
sentence-transformers==2.2.2

# Database connectivity
psycopg2-binary==2.9.9

# Configuration and utilities
python-dotenv==1.0.0
pydantic==2.5.2
python-multipart==0.0.6
httpx==0.25.2

# Numerical operations
numpy==1.24.3

# Optional for enhanced performance
orjson==3.9.10
python-multipart==0.0.6
